requirements.txt
gradio==5.9.1
torch
torchvision
opencv-python-headless
pillow
numpy
transformers>=4.39.1
loadimg>=0.1.1
